+++ mouse/events.go
@@ -4,15 +4,15 @@ import (
 	"github.com/veandco/go-sdl2/sdl"
 )
 
-type ButtonPressCB func(x, y float64, button Button)
-type ButtonReleaseCB func(x, y float64, button Button)
-type MoveCB func(x, y, dx, dy float64)
+type ButtonPressCB func(x, y float32, button Button)
+type ButtonReleaseCB func(x, y float32, button Button)
+type MoveCB func(x, y, dx, dy float32)
 type FocusCb func(has_focus bool)
 
 var (
-	button_press_default   ButtonPressCB   = func(x, y float64, button Button) {}
-	button_release_default ButtonReleaseCB = func(x, y float64, button Button) {}
-	move_default           MoveCB          = func(x, y, dx, dy float64) {}
+	button_press_default   ButtonPressCB   = func(x, y float32, button Button) {}
+	button_release_default ButtonReleaseCB = func(x, y float32, button Button) {}
+	move_default           MoveCB          = func(x, y, dx, dy float32) {}
 	focus_default          FocusCb         = func(has_focus bool) {}
 
 	button_press_cb   = button_press_default
@@ -24,13 +24,13 @@ var (
 func Delegate(event sdl.Event) {
 	switch e := event.(type) {
 	case *sdl.MouseMotionEvent:
-		move_cb(float64(e.X), float64(e.Y), float64(e.XRel), float64(e.YRel))
+		move_cb(float32(e.X), float32(e.Y), float32(e.XRel), float32(e.YRel))
 	case *sdl.MouseButtonEvent:
 		switch e.Type {
 		case sdl.MOUSEBUTTONDOWN:
-			button_press_cb(float64(e.X), float64(e.Y), Button(e.Button))
+			button_press_cb(float32(e.X), float32(e.Y), Button(e.Button))
 		case sdl.MOUSEBUTTONUP:
-			button_release_cb(float64(e.X), float64(e.Y), Button(e.Button))
+			button_release_cb(float32(e.X), float32(e.Y), Button(e.Button))
 		}
 	case *sdl.WindowEvent:
 		focus_cb(e.Type == sdl.WINDOWEVENT_ENTER)
